// FRONT END/moduloEnvio/src/app/models/envio.ts
export class Envio {
  idenvio: number;
  estado: string;
  tipo_dimension: number;
  idPedido: number;
  tipo_peso: number;
  alto: number;
  direccion: string;
  fechaEntrega: Date;
  fondo: number;
  largo: number;
  metodoEnvio: string;
  nota: string;
  numeroSeguimiento: number;
  peso: number;
  nombreCliente ? : string;
}